feat: Add food consumption tracking functionality
This commit adds functionality to track food consumption in the `index` view of the `myapp` application. Users can now submit a form with the name of the food they consumed, and it will be saved in the `Consume` model along with the user information. The `index` view has been updated to display all food items and the food items consumed by the user.